Move synchronizer to own file
